Python script now outputs data to a file.
A file is created at $MEMBERWORK/.wraprun/tmp${PBS_JOBID}_${HASH}.wraprun

The HASH is a random string produced by the python tempfile package.

The directory is created if it doesn't exist with mode 700 and the files are
created with mode 600 owned by the user with user's primary group.

The path to the file is stored in the environment variable WRAPRUN_FILE.