Generalized group argument parsing.
Exectuables can now be split into multiple colors by specifying comma-separated
values to certain argument flags. Flags that support color splitting are:

  Aprun flags:
  Number of PEs `-n`

  Wraprun flags:
  Path cd `--w-cd`

Future arguments can be added by subclassing the FirstOccurance(argparse.Action)
class with the attribute "splits = True" and overriding the ns_format and
cli_format methods as necessary.

Default values for splitting arguments must be lists of strings.

WARNING:
`-n` is required and wraprun flags (`--w-X`) are unique and unlikely to conflict
with executable arguments. The first instance of any group args searched for is
assumed to be a wraprun/aprun arg and any duplicated argument is assumed to have
been caught after the executable and are moved to the end of the group arg list.

Until there is a way to verify the executable in the arglist, or all aprun args
are searched for with known numbers of expected values, there is no way to
sanity check against flag collisions between aprun and the executable.