BROKEN - Initial commit of API features.
Added skeletal API interface and initial restructuring to read input from
configuration files. Two new files added that will eventually replace
wraprun.py.

The current plan is to create a new module to replace the CLI. The dedicated CLI
module will be a translation layer that uses the new API as a backend.

All argument options will be listed in their own objects (currently in a set of
dicts, but possibly in the future their own classes) that contain the argument
strings, order, and parser details for that argument. ArgumentParser and API
elements can use the same objects such that new options have only one point of
access.