Limits maximum number of task groups to 2048.
To protect system stability, maximum number of task groups is limited to 2048
':'-separated executable groups. This caps the number of allocations per unique
ALPS Reservation-Application ID pair. ALPS chokes when the number of allocations
in an application is large, but no practical maximum is discussed in the Cray
documentation. As such, this number is arbitrarily chosen and is being reduced
as administrators report issues.